backend/requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
requests
whitenoise
gunicorn
pillow
resend
